+++ ui/linux-qt/src/MainWindow.cpp
@@ -1552,6 +1552,7 @@ MainWindow::MainWindow(tesseract::AccountManager& account_manager, QWidget* pare
                 };
             mainApp_->open_camera_overlay();
         };
+        sh.on_location = [this] { send_current_location_(current_room_id_); };
         slash_controller_ =
             std::make_unique<tesseract::views::SlashCommandController>(
                 roomTextArea_, slash_popup_widget_, std::move(sh));